dispense_demo3/pyproject.toml
[project]
name = "dispense-demo3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "psycopg2>=2.9.10",
    "python-dotenv>=1.0.1"
]
